inno_afisha/requirements.txt
Django==4.0
django-cors-headers==3.10.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==22.1
psycopg2-binary==2.9.3